+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 # 7.0.2 [Firefox: Popup Menu fix]
 
 - Popup menu can now change colors on pages other than _New Tab_ pages in **mix mode**
+- Adds bounce animation to doki theme icon
 
 # 7.0.1 [Rename Repository]